1차 홈수정

+++ src/main/java/com/dbjava/javabang/respository/AccommodationRepository.java
@@ -4,7 +4,7 @@ import com.dbjava.javabang.domain.entity.Accommodation;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccommodationRepository extends JpaRepository<Accommodation, Integer> {
+public interface AccommodationRepository extends JpaRepository<Accommodation, Long> {
 
   List<Accommodation> findAllByUserId(Long userId);
 }
